% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bSplineScores.R
\name{bSplineScores}
\alias{bSplineScores}
\title{Get scores based on fixed B-spline basis.}
\usage{
bSplineScores(ds, df)
}
\arguments{
\item{ds}{Data frame as generated by \code{\link{generateData}}.}

\item{df}{Number of degrees of freedom to use.}
}
\value{
List containing \code{ds}, a data frame containing the scores, and
\code{aic1} and \code{aic2} the AIC values for the mixed effects model for
y1 an y2, respectively, to be used for downstream model selection.
}
\description{
Get scores based on fixed B-spline basis in a linear mixed effects model.
}

