% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpca.R
\name{fpca}
\alias{fpca}
\title{Fit functional principal components analysis}
\usage{
fpca(ds, ycols, tcol, idcol, options)
}
\arguments{
\item{ds}{Data frame containing data for analysis, one row per observation.}

\item{ycols}{Character vector identifying columns of \code{ds} corresponding to longitudinal outcomes.}

\item{tcol}{Character vector identifying the column of \code{ds} corresponding to time.}

\item{idcol}{Character vector identifying the column of \code{ds} corresponding to patient ID.}

\item{options}{List of options to pass to \code{FPCA}, see documentation at \code{\link[fdapace]{FPCA}}.}
}
\value{
A list containing a data frame of FPCA scores and IDs, a data frame
   containing eigenfunction evaluations at observed time points and IDs,
   and a list of FPCA fits.
}
\description{
Fit FPCA and return scores matched to IDs and return eigenfunctions
matched to IDs.
}

