% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpme.R
\name{fpme}
\alias{fpme}
\title{Get re-fitted scores based on FPCA basis.}
\usage{
fpme(fpca, df)
}
\arguments{
\item{fpca}{Object from FPCA fitting.}

\item{df}{Number of degrees of freedom to use.}
}
\value{
A data frame containing the scores.
}
\description{
Get scores based on a linear mixed effects model using the FPCA
eigenfunctions as a basis for time.
}

