% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpvc.R
\name{fpvc}
\alias{fpvc}
\title{Perform testing of summarized longitudinal trajectory.}
\usage{
fpvc(data, xicols, zcols, xcols = NULL, model)
}
\arguments{
\item{data}{Data frame containing scores summarizing longitudinal trajectories,
genetic markers, and any covariates to adjust for, one row per
individual.}

\item{xicols}{Character vector indicating columns of \code{data} that
correspond to the scores summarizing longitudinal trajectory.}

\item{zcols}{Character vector indicating columns of \code{data} that
correspond to the genetic markers.}

\item{xcols}{Character vector indicating columns of \code{data} that
correspond to the covariates to adjust for.}
}
\value{
A p-value.
}
\description{
Perform testing of summarized longitudinal trajectory.
}

