% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateData.R
\name{generateData}
\alias{generateData}
\title{Generate data for simulation.}
\usage{
generateData(n, nt, maf, beta, TT, alpha, gamma)
}
\arguments{
\item{n}{Number of samples.}

\item{nt}{Average number of time points.}

\item{maf}{Minor allele frequency of SNP.}

\item{beta}{Effect size.}

\item{TT}{Length of study period.}

\item{alpha}{Parameter controlling the linearity of genetic effect, between 0
(nonlinear effect) and 1 (linear effect).}

\item{beta}{Parameter controlling the complexity of the mean process between
0 (simple mean process) and 1 (complex mean process)}
}
\description{
Generate data with varying amounts of nonlinearity and complexity consistent
with the simulations for "Functional principle variance component testing
for a genetic association study of HIV progression."
}

