% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreTest.R
\name{scoreTest}
\alias{scoreTest}
\title{Variance component score test.}
\usage{
scoreTest(y, z, x = NULL, model)
}
\arguments{
\item{y}{The outcomes of interest, n x m matrix.}

\item{z}{The predictors of interest, n x p matrix.}

\item{x}{Covariates to adjust for, n x q matrix.}

\item{model}{Model used to adjust for \code{x}, currently only proportional
odds ('po'), binomial, logistic, and probit supported.}
}
\value{
p-value of the test
}
\description{
Tests for an association between multivariate \code{y} and multivariate \code{z} conditional
on multivariate \code{x}.
}
\examples{
n <- 1000
p <- 2
q <- 5
m <- 3
x <- rnorm(n*q) \%>\% matrix(n, q)
pi <- pnorm(x)
z <- rbinom(n*p, size = 1, prob = pi) \%>\% matrix(n, p)
mu <- rowMeans(x) + z[,1]/10
y <- rnorm(n*m, mean = mu) \%>\% matrix(n, m)
scoreTest(y = y, z = z, x = x, model = 'probit')

}

